refactor: extract the AdhocRunner lane from DispatchOperations
First launching lane peeled, now that the three seams are in place. The
ad-hoc lane (sail agent run --task) is a thin lane over them: reserve the
whole container through RunReservation, launch through RunLauncher, release
the reservation through RunReservation if the launch fails.

Move startAdhoc and its helpers (reserveAdhocRun, adhocWorkDir,
adhocFullPermissions, prepare) into AdhocRunner over five collaborators
(projects, runLauncher, runReservation, runStore, listener). DispatchOperations
keeps startAdhoc as a two-line delegator so the public surface and callers
(AgentSweepCommand, RunCommand) are unchanged.

Behavior-preserving — AdhocDispatchTest passes unchanged and covers every
AdhocRunner branch (dry run, live launch, no-store refusal, prepare failure,
full/limited permissions), so the new api class meets the 100% coverage gate
with no exclude and no new test. mvn clean verify green.
